+++ src/main/java/com/keepgoing/keepserver/global/file/mapper/RowMapper.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.global.file;
+package com.keepgoing.keepserver.global.file.mapper;
 
 import org.apache.poi.ss.usermodel.Row;
